Disable ld.so.cache support
The ld.so.cache file never exists in a NaCl manifest and there
is no benefit to using such a file in the NaCl context.  When
using sel_ldr -a or the environment variable equivalent on a
Linux host, searching for this file can hit the native host
file and have unintended effects on library search.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=3316
TEST= toolchain trybot

R=bradnelson@google.com

Review URL: https://codereview.chromium.org/12340013