Allow use of alternative 'main' symbol.
If the __nacl_main symbol is found at runtime
them use this in place of the normal 'main'.

The idea is that if/when libppapi_simple.a is
linked into an executable __nacl_main will be
called as the startup symbol and then the real
main symbol can be called later on (e.g. on a
separate thread).

BUG=246246
R=mcgrathr@chromium.org

Review URL: https://codereview.chromium.org/16514007